removing getWriteContext() test methods
